+++ lib/config/rate-limits.ts
@@ -1,15 +1,19 @@
-export interface MessageRateLimits {
+export interface RateLimits {
   highPrivmsgLimits: number;
   lowPrivmsgLimits: number;
 
   // whispersPerSecond: number;
   // whispersPerMinute: number;
   // whisperTargetsPerDay: number;
+
+  privmsgInMs: number;
+
+  joinLimits: number;
 }
 
 export type PresetKeys = "default" | "knownBot" | "verifiedBot";
 
-export const messageRateLimitPresets: Record<PresetKeys, MessageRateLimits> = {
+export const rateLimitPresets: Record<PresetKeys, RateLimits> = {
   default: {
     highPrivmsgLimits: 100,
     lowPrivmsgLimits: 20,
@@ -17,6 +21,10 @@ export const messageRateLimitPresets: Record<PresetKeys, MessageRateLimits> = {
     // whispersPerSecond: 3,
     // whispersPerMinute: 100,
     // whisperTargetsPerDay: 40
+
+    privmsgInMs: 35 * 1000,
+
+    joinLimits: 20,
   },
   knownBot: {
     highPrivmsgLimits: 100,
@@ -25,6 +33,10 @@ export const messageRateLimitPresets: Record<PresetKeys, MessageRateLimits> = {
     // whispersPerSecond: 10,
     // whispersPerMinute: 200,
     // whisperTargetsPerDay: 500
+
+    privmsgInMs: 35 * 1000,
+
+    joinLimits: 20,
   },
   verifiedBot: {
     highPrivmsgLimits: 7500,
@@ -33,5 +45,9 @@ export const messageRateLimitPresets: Record<PresetKeys, MessageRateLimits> = {
     // whispersPerSecond: 20,
     // whispersPerMinute: 1200,
     // whisperTargetsPerDay: 100000
+
+    privmsgInMs: 35 * 1000,
+
+    joinLimits: 2000,
   },
 };