+++ lib/mixins/ratelimiters/index.ts
@@ -1,4 +1,5 @@
 export * from "./connection";
+export * from "./join";
 export * from "./privmsg";
 export * from "./slow-mode";
 export * from "./utils";