+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unversioned
 
+- Added JoinRateLimiter mixin. (#11)
 - Dev: Added TypeScript IDE type hints for event names. (#12)
 
 ## v6.0.0